move flight reference fixtures package-side
