+++ .changeset/clean-flights-reference-fixtures.md
@@ -1,0 +1,5 @@
+---
+"@voyant-travel/flights": minor
+---
+
+Own the curated airline, airport, and aircraft reference fixture behind the supported `@voyant-travel/flights/reference/fixtures` API.
