+++ packages/flights/src/reference/fixtures/types.ts
@@ -1,0 +1,3 @@
+import type { referenceAirports } from "../local-postgres.js"
+
+export type AirportFixtureRow = typeof referenceAirports.$inferInsert
