--- starters/operator/scripts/seed-flights-reference.ts
@@ -1,8 +1,0 @@
-/**
- * Operator flights-reference seed helpers. Public exports are preserved while
- * curated fixture rows live in focused airline, airport, and aircraft modules.
- */
-
-export { seedAircraft } from "./seed-flights-reference-aircraft"
-export { seedAirlines } from "./seed-flights-reference-airlines"
-export { seedAirports } from "./seed-flights-reference-airports"
